requirements.txt
Django==3.2.15
psycopg2_binary==2.9.3
python-dotenv==0.15.0
gunicorn==20.1.0
dj-static==0.0.6
djangorestframework==3.14.0
django-cors-headers==3.13.0
requests==2.25.1
geopy==2.2.0
djangorestframework-simplejwt==5.2.0
flake8
uWSGI==2.0.20
